+++ CHANGELOG.md
@@ -9,12 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+## [v3.0.0-beta.2] - 2021-08-11
+### Changed
+- Pull incoming changes from [v2.1.0]
+- Update `actix-web` to `4.0.0-beta.8`
+- Remove RefCell in middleware [#11](https://github.com/DDtKey/actix-web-grants/pull/11)
+
 ## [v2.1.0] - 2021-08-11
 ### Added
 - Add support for ABAC-like model to procedural macro [#14](https://github.com/DDtKey/actix-web-grants/issues/14)
 
 ### Changed
 - Change Arc to Rc in middleware [#12](https://github.com/DDtKey/actix-web-grants/pull/12)
+
+## [v3.0.0-beta.1] - 2021-04-07
+### Changed
+- Update `actix-web` to `4.0.0-beta.5`
 
 ## [v2.0.1] - 2021-03-10
 ### Changed
@@ -44,4 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v0.1.6]: https://crates.io/crates/actix-web-grants/0.1.6
 [v1.0.0]: https://crates.io/crates/actix-web-grants/1.0.0
 [v2.0.0]: https://crates.io/crates/actix-web-grants/2.0.0
+[v2.0.1]: https://crates.io/crates/actix-web-grants/2.0.1
 [v2.1.0]: https://crates.io/crates/actix-web-grants/2.1.0
+[v3.0.0-beta.1]: https://crates.io/crates/actix-web-grants/v3.0.0-beta.1
+[v3.0.0-beta.2]: https://crates.io/crates/actix-web-grants/v3.0.0-beta.2